Reduce calls to Task.cancel() when possible
An `await Task.cancel()` might result in a context switch even if the
task has been terminated; thus it might be a good idea to reduce the
number of unnecessary Task.cancel() calls.  I agree that this might be a
premature optimization, but since it's pretty trivial to do, why not?

I guess the reason that Task.cancel() does a context switch even on
terminated tasks is for consistency: there will always be a context
switch when you call Task.cancel() regardless the task state.