Make garage.cli sub-commands interface consistent
Prior to this change, sub-commands have to be wrapped in
call_sub_command_entry_point, and as a result, you have to call sub-
commands through the wrapper at `args.subcmd_dest` and cannot call them
directly, which is inconsistent with the way you call the main command.

This change makes the command objects detect whether they are the main-
or a sub- command, and behaves accordingly.