Rename a foreman feature "parametric dependency"
While foreman has this feature for some time, its name was a misnomer as
this feature does not really configure something.

I think it should be better named parametric dependency because what it
actually does is passing parameters to the dependent rule's execution
environment.  Also the name parametric should suggest that it generates
edge per "instantiation" in the extended dependency graph.