Fix a tasklet race condition and clarify design limitation
The race condition is like this: after `task()` returns but before the
tasklet notifies the task queue, another tasklet gets a task from the
queue, making the queue be empty.  Then the first tasklet notifies the
queue it finishes its task, and the queue sees that it is empty now (but
it doesn't know that the second tasklet is still running and may spawn
more tasks) and closes itself.

Contrary to the standard library's Queue class which keeps a separate
counter of "queue length" that is incremented when you call `put()` and
decremented when you call a special method `task_done()`, we track the
number of running tasklets.  I'm not sure which approach is better.