Replace AsyncStub with StubAdapter
Because AsyncStub alters function's semantics, it is not a true subclass
of Stub in the sense of Liskov Substitution Principle, which makes it
hard when we need both synchronous and asynchronous stub.

Since all AsyncStub does is wrapping result in a FutureAdapter object,
we could instead use an adapter rather than subclassing Stub.