Update docs on merging "config data" volume into image
In our earlier release practice, we strictly separated where code and
data were stored; containerized images only stored code, and all data
must be separated from images and were stored in data volumes.

The benefit of this practice is that it reduced the number of images
that only differ in the config data they contain (because these config
data are stored separately).  But this benefit is outweighed by the
burden on ops to maintain lots of tiny "config data" volumes.

So we decide to merge these config data volumes into images; the total
number of build artifacts (images plus volumes) will be smaller, and ops
will no longer be maintaining the map of image version to config data
volume version (which is a fragile process).