Downgrade JDK to 8u172
This temporarily works around this build issue:
```
java.lang.ClassNotFoundException: javax.xml.bind.ValidationEventHandler
```

JAXB API is part of Java EE, and is not contained in the default class
path in Java SE 9.  To properly fix this, one should add JAXB to Gradle
build dependencies, but I am no expert in Gradle, and am not eager to
upgrade to Java 9; so let's stay in Java 8 for now.