Allocate port numbers per instance as well
Previously, port numbers are allocated on a per-pod basis.  Since we
promote the status of unit instance to the same level of pod name and
version, port numbers should be allocated per each combination of pod,
version, and instance.

NOTE: After this change, the only resource that is not allocated way is
volume.  (Which might be beneficial because instances may share volume
data; so let's keep it this way for now.)

Also, updating Ports class that:
* Allocated ports are guaranteed to be unique.
* Statically assigned ports can be assigned to multiple instances.