Clarify instance-related operations
These are operations related to systemd unit instances: start, stop,
enable, disable, is-started, and is-enabled.  Their behavior was unclear
because "unit instance" was an afterthought.

Now, for all of them, the default behavior is to operate on default unit
instances (as specified in the pod file).  You may use --instance or
--instance-all command-line argument to change its default behavior.

As for undeploy, it stops and disables all instances, since leaving some
instance not stopped or not disabled after a pod is undeployed does not
make sense.