Change queue close interface
When ``close`` is called at multiple places, in the old interface, the
first call site determines whether the queued items are dropped, but in
the new interface, the first call site with ``graceful=False`` makes
that decision.

The new interface gives every call sites a chance to drop queued items,
and should be friendlier to framework codes.