Switch to epoll edge trigger
With level trigger (default), we have to disarm a fd after a task
blocked on the fd is awaken or cancelled, or else the event loop could
be trapped in busy polling the fd.

Switching to edge trigger simplifies the internals of the event loop,
and should be (slightly?) more performant.