Do not pin base image version
rkt pins stage1 image version to the same version of rkt.  This makes
sense for rkt because they define a clear interface of stage1 that rkt
relies on, and this version-pinning provides stability.

On the other hand, we are still evolving our container runtime design,
and we would prefer flexibility of changing base image design over
stability of base image interface.