Work around issue of different sets of parameters
When foreman is launched with different rules, it might load different
sets of build files, and thus different sets of parameters.  This causes
trouble for foreman wrapper scripts that add default parameters, as
these parameters might not be loaded under certain input rules.  We work
around this issue by adding a dummy rule and making most rules depend on
this dummy rule (so that build files are always loaded).