Rewrite shipyard2.rules.{images,pods}
We used to separately implement functionalities in the builder script,
and foreman build rules.  This complicates the internal architecture,
and duplicates code (specifically, parameters have to be tossed around).

This change is the first step that implement the entire build and
release process in build rules.  The builder and the release script will
be merely launching foreman (this part will be done in the following
changes).