Make commands idempotent
It is important to note that a command is idempotent does not mean it
will not err out; for example, commands will still err out if they
cannot acquire relevant file locks.

For define/import/install command, idempotent means that they will
behaves like an "upsert" operation.

For remove/uninstall command, idempotent means that they will be no-op
if the target does not exist.