Remove calling context manager of actor object
While this idea looks neat, it has a flaw that if __enter__ raises, the
actor will crash, which is quite hard to detect; you either:

* Wait on stub.future for a short period of time, and assume __enter__
  succeeds after the timeout.
* Call a dummy method with a timeout, and assume __enter__ fails after
  the timeout.

The __enter__ idea does not look that neat if we need one of these
workarounds every time with __enter__.  Instead, it is probably better
that the actor initialization is so simple that it never fails.

If a user wants to do complex initialization, it can do that right after
stub creation by calling actor's initialization.  Since at this point,
the actor's stub is not shared with anyone else yet, it is guaranteed
that this initialization is the first method that the actor receives.

If a user wants to do cleanup, it can call actor's cleanup, and before
waiting for the result, call stub.close.  NOTE: There is no guarantee
that no other method calls sneak in between cleanup and close (albeit
the chances are low).  The actor should enter into a "closed" state
after it processed the cleanup call, rejecting all future calls; this is
similar to the behavior of a file object after it is closed.