Add the assignment names to token file format
Before this change, tokens from the same group are indistinguishable
(because they are share the same name).  This change adds names to
assignments, which can be used to distinguish tokens from the same
group.

Pod deploy instruction already implemented this, called "local alias",
for the same purpose.

NOTE: This change breaks backward compatibility of token file format.
(We could make this change in a backward compatible way, but meh.)