Rewrite kernel I/O poller
The old poller abstraction was leaky, and the interface was complex.
Even worse, we suspect that it fails to notify the kernel when a file
descriptor becomes ready, probably causing tasks to be retained longer,
and the process running out of file descriptors.  Hopefully the new
poller will fix these issues.