Change circuit breaker YELLOW state to err out on concurrent requests
We use GREEN to indicate that the data source is up, and RED when it is
down.  The YELLOW state is when the breaker is trying to detect whether
the data source is up again.  So it make more sense to make the breaker
errs out all other concurrent requests, as the data source should still
be treated as "down" in this state.