Test how multiprocessing interacts with signal
It turns out that CPython behaves differently when Ctrl-C is pressed vs
SIGINT is delivered (to the main process or to all processes).  This is
annoying, but we will have to remember that Ctrl-C might not be able
stop a Python script that uses multiprocessing.

Also, the signal handling strategy of the multiprocessing module, while
looks weird, turns out to be "correct", and our monkey-patching module
g1.bases.multiprocessings is wrong.